# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank management system
  description: api for managing user's cards
  contact:
    name: Panasik Uladzislau
    url: https://github.com/Vladislav3421730
    email: panasikvladislav1@gmail.com
  version: 1.0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /api/v1/transfer:
    post:
      tags:
        - Transfer Management
      summary: Transfer funds
      description: Transfer money from one card to another
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequestDto'
        required: true
      responses:
        '200':
          description: Transfer was executed successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TransferResponseDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/transfer/withdrawal:
    post:
      tags:
        - Transfer Management
      summary: Withdraw money
      description: Initiate a withdrawal transaction for the current user
      operationId: withdrawal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawalRequestDto'
        required: true
      responses:
        '200':
          description: Transfer was executed successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/WithdrawalResponseDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/transfer/recharge:
    post:
      tags:
        - Transfer Management
      summary: Recharge card
      description: Recharge a card with a specified amount for the current user
      operationId: recharge
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RechargeRequestDto'
        required: true
      responses:
        '200':
          description: Transfer was executed successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RechargeResponseDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/users/{id}/ban:
    patch:
      tags:
        - Users Management
      summary: Ban or unban user
      description: Updates the banned status of a user
      operationId: banUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BannedRequestDto'
        required: true
      responses:
        '204':
          description: User ban status updated successfully
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldErrorDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/users:
    get:
      tags:
        - Users Management
      summary: Find all users
      operationId: findAll_1
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: List of users returned successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
  /api/v1/users/{id}:
    get:
      tags:
        - Users Management
      summary: Get user by ID
      description: Retrieve user information by their unique identifier
      operationId: findById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
    delete:
      tags:
        - Users Management
      summary: Delete user
      description: Deletes a user by their ID
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User successfully deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
  /api/v1/cards/{id}:
    get:
      tags:
        - Card Management
      summary: Get card by ID
      description: Retrieve card information by their unique identifier
      operationId: findById_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: card found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
    delete:
      tags:
        - Card Management
      summary: Delete card
      description: Deletes a card by its ID
      operationId: deleteCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Card successfully deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppErrorDto'
      security:
        - BearerAuthentication: []
...
components:
  schemas:
...
    FieldErrorDto:
      type: object
      description: DTO for errors with validation
      properties:
        code:
          type: integer
          format: int32
          description: Error's code
          example: 400
        timestamp:
          type: string
          description: Timestamp of when the error occurred
          example: '2025-02-07 14:30:00'
        errors:
          type: object
          additionalProperties:
            type: string
          description: Error map describing the validation
    AppErrorDto:
      type: object
      description: DTO for errors
      properties:
        code:
          type: integer
          format: int32
          description: Error's code
          example: 400
        timestamp:
          type: string
          description: Timestamp of when the error occurred
          example: '2025-02-07 14:30:00'
        message:
          type: string
          description: Error message describing the problem
          example: Invalid request
...
    TransferRequestDto:
      type: object
      description: DTO for transferring funds between two cards
      properties:
        number:
          type: string
          description: Card number from which funds will be transferred
          example: 1234567890123456
        targetNumber:
          type: string
          description: Card number to which funds will be transferred
          example: 6543210987654321
        amount:
          type: number
          description: Amount to be transferred
          example: 100
          minimum: 5
    TransferResponseDto:
      type: object
      description: Response DTO for transfer transaction details
      properties:
        cardId:
          type: string
          format: uuid
          description: Card ID associated with the transfer
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        number:
          type: string
          description: Card number associated with the transfer
          example: 1234567890123456
        targetCardId:
          type: string
          format: uuid
          description: Target card ID for the transfer
          example: b34f1b88-70be-413a-bf34-f3a1fa6f3a24
        targetNumber:
          type: string
          description: Target card number for the transfer
          example: 9876543210987654
        transferAmount:
          type: number
          description: Amount transferred from the card
          example: 50
        transferTime:
          type: string
          format: date-time
          description: Timestamp of the transfer transaction
          example: '2025-04-14 14:30:00'
        balance:
          type: number
          description: New balance after the transfer
          example: 150
        userId:
          type: string
          format: uuid
          description: User ID associated with the card
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
    WithdrawalRequestDto:
      type: object
      description: DTO for withdrawing funds from a card
      properties:
        number:
          type: string
          description: Card number from which funds will be withdrawn
          example: 1234567890123456
        amount:
          type: number
          description: Amount to be withdrawn
          example: 50
          minimum: 5
    WithdrawalResponseDto:
      type: object
      description: Response DTO for withdrawal transaction details
      properties:
        cardId:
          type: string
          format: uuid
          description: Card ID associated with the withdrawal
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        number:
          type: string
          description: Card number associated with the withdrawal
          example: 1234567890123456
        withdrawalAmount:
          type: number
          description: Amount withdrawn from the card
          example: 50
        transferTime:
          type: string
          format: date-time
          description: Timestamp of the withdrawal transaction
          example: '2025-04-14 14:30:00'
        remainingBalance:
          type: number
          description: Remaining balance after the withdrawal
          example: 150
        userId:
          type: string
          format: uuid
          description: User ID associated with the card
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
    RechargeRequestDto:
      type: object
      description: DTO for recharging a card
      properties:
        number:
          type: string
          description: Card number to recharge
          example: 1234
        amount:
          type: number
          description: Recharge amount
          example: 100
          minimum: 5
    RechargeResponseDto:
      type: object
      description: Response DTO for recharge transaction details
      properties:
        cardId:
          type: string
          format: uuid
          description: Card ID associated with the recharge
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        number:
          type: string
          description: Card number associated with the recharge
          example: 1234567890123456
        rechargeAmount:
          type: number
          description: Amount recharged to the card
          example: 100
        transferTime:
          type: string
          format: date-time
          description: Timestamp of the recharge transaction
          example: '2025-04-14 14:30:00'
        balance:
          type: number
          description: New balance after the recharge
          example: 200
        userId:
          type: string
          format: uuid
          description: User ID associated with the card
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
...
    CardDto:
      type: object
      description: DTO representing a card with details such as number, balance, and expiration date.
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the card
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        number:
          type: string
          description: Card number
          example: '**** **** **** 3456'
        status:
          type: string
          description: Status of the card
          enum:
            - ACTIVE
            - INACTIVE
            - BLOCKED
          example: ACTIVE
        balance:
          type: number
          description: Current balance on the card
          example: 1000.5
        expirationDate:
          type: string
          format: date
          description: Expiration date of the card
          example: '2025-12-31'
        userId:
          type: string
          format: uuid
          description: User ID associated with the card
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
...
    BannedRequestDto:
      type: object
      description: Request to change banned status of a user or card
      properties:
        banned:
          type: boolean
          description: Indicates whether the entity should be banned (true) or unbanned (false)
          example: true
      required:
        - banned
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModel:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          $ref: '#/components/schemas/PageMetadata'
    UserDto:
      type: object
      description: DTO representing a user and their associated cards.
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        username:
          type: string
          description: User login
          example: user
        firstName:
          type: string
          description: First name of the user
          example: John
        lastName:
          type: string
          description: Last name of the user
          example: Doe
        cards:
          type: array
          description: List of cards associated with the user
          items:
            $ref: '#/components/schemas/CardDto'
        isBan:
          type: boolean
          description: Ban status of the user
          example: false
  securitySchemes:
    BearerAuthentication:
      type: http
      scheme: bearer
      bearerFormat: JWT